+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/rootcause/RCAFrameworkExecutionResult.java
@@ -1,0 +1,41 @@
+package com.linkedin.thirdeye.rootcause;
+
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+
+
+/**
+ * Container object for framework execution results. Holds the results aggregated across all
+ * pipeline executions as well as the results for each individual pipeline (keyed by pipeline name).
+ *
+ */
+public final class RCAFrameworkExecutionResult {
+  private final Set<Entity> results;
+  private final Map<String, PipelineResult> pipelineResults;
+
+  public RCAFrameworkExecutionResult(Set<? extends Entity> results, Map<String, PipelineResult> pipelineResults) {
+    this.results = new HashSet<>(results);
+    this.pipelineResults = pipelineResults;
+  }
+
+  /**
+   * Returns the flattened results of a framework execution (i.e. the results of the
+   * {@code RCAFramework.OUTPUT} pipeline).
+   *
+   * @return flattened framework execution results
+   */
+  public Set<Entity> getResults() {
+    return results;
+  }
+
+  /**
+   * Returns a map of sets of results as generated by each individual pipeline during execution.
+   * The map is keyed by pipeline name.
+   *
+   * @return map of pipeline results keyed by pipeline name
+   */
+  public Map<String, PipelineResult> getPipelineResults() {
+    return pipelineResults;
+  }
+}
